Refactored Fix Naming into it's specific file and separated camel case detection/replacement logic into a util file
